Organize CSS into directories with their respective components, and update README
