+++ src/app/components/Contact/Contact.css
@@ -1,0 +1,224 @@
+#contact {
+  margin: 0;
+  position: relative;
+  transform-style: preserve-3d;
+  width: 100%;
+  padding-top: 5%;
+}
+
+#contact-contents {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin: 0 20%;
+}
+
+#contact-h3 {
+  display: flex;
+  align-items: center;
+  margin: 0 20%;
+}
+
+#email-icon {
+  margin-left: 2%;
+}
+
+.contact-form {
+  margin: 0 5% 5%;
+  padding: 1% 5%;
+  border-top: 1px solid var(--brown);
+  background-color: var(--dark-brown-1);
+  display: flex;
+  flex-direction: column;
+  border-radius: 5px;
+  min-width: 50%;
+  max-width: 50%;
+}
+
+.form-div {
+  margin: 3% 0 1%;
+}
+
+.form-label {
+  padding: 5% 0;
+}
+
+.form-input,
+textarea {
+  background-color: transparent;
+  color: var(--white);
+  font-size: clamp(0.7rem, 1.1vw, 1.3rem);
+  padding: 0.6em;
+  border-radius: 10px;
+  border: 1px solid var(--brown);
+  font-family: inherit;
+  letter-spacing: 0.1em;
+  width: 95%;
+}
+
+textarea {
+  height: 5vw;
+  resize: none;
+}
+
+.form-input:focus,
+textarea:focus {
+  outline: none;
+  border-color: var(--beige-2);
+}
+
+.form-submit {
+  border: 1px solid var(--brown);
+  background-color: var(--brown-2);
+  width: 6em;
+  font-family: inherit;
+  color: var(--white);
+  font-size: inherit;
+  margin: 5% 0;
+  padding: 2%;
+  border-radius: 10px;
+  letter-spacing: inherit;
+}
+
+.form-submit:hover {
+  border-color: var(--beige-2);
+  background-color: var(--brown-3);
+  cursor: pointer;
+}
+
+.submit-confirmation {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  font-size: clamp(0.8rem, 1.15vw, 3rem);
+  margin-top: 2%;
+}
+
+.submit-icon {
+  font-size: clamp(0.8rem, 1.8vw, 3rem);
+}
+
+.confirmation-msg {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.confirmation-txt {
+  margin: 0;
+}
+
+#email-message {
+  font-size: clamp(0.8rem, 1vw, 3rem);
+  color: var(--beige-3);
+  z-index: 3;
+  margin: 0;
+}
+
+.place-holder {
+  visibility: hidden;
+}
+
+.parallax-bot-boops {
+  width: 100%;
+  height: 100%;
+  position: absolute;
+  top: -35%;
+  z-index: -1;
+}
+
+#parallax-bot-boops {
+  -ms-transform: translateZ(-8px) translateY(-80px) scale(1.5);
+  -webkit-transform: translateZ(-8px) translateY(-80px) scale(1.5);
+  -moz-transform: translateZ(-8px) translateY(-80px) scale(1.5);
+  -o-transform: translateZ(-8px) translateY(-80px) scale(1.5);
+  transform: translateZ(-8px) translateY(-80px) scale(1.5);
+}
+
+#parallax-bot-tendrils1 {
+  -ms-transform: translateZ(-12px) scale(1.8);
+  -webkit-transform: translateZ(-12px) scale(1.8);
+  -moz-transform: translateZ(-12px) scale(1.8);
+  -o-transform: translateZ(-12px) scale(1.8);
+  transform: translateZ(-12px) scale(1.8);
+}
+
+#parallax-bot-tendrils2 {
+  -ms-transform: translateZ(-16px) scale(2.1);
+  -webkit-transform: translateZ(-16px) scale(2.1);
+  -moz-transform: translateZ(-16px) scale(2.1);
+  -o-transform: translateZ(-16px) scale(2.1);
+  transform: translateZ(-16px) scale(2.1);
+}
+
+#credit-text {
+  font-size: clamp(0.8rem, 1vw, 1rem);
+  color: var(--beige-3);
+  position: absolute;
+  bottom: -20%;
+  padding: 0;
+  text-align: center;
+  width: 100%;
+}
+
+
+/* ------ Media Queries ------- */
+
+
+@media screen and (max-width: 882px) {
+  #email-message {
+    top: 43%;
+  }
+}
+
+@media screen and (max-width: 768px) {
+  #email-message {
+    top: 37%;
+  }
+
+  #contact {
+    margin-top: 5%;
+  }
+
+  #contact-contents {
+    margin: 0 5%;
+  }
+
+  .submit-confirmation {
+    top: 8%;
+    left: 25%;
+  }
+
+  .form-input,
+  textarea {
+    width: 90%;
+    font-size: clamp(1rem, 0.2vw, 1.3rem);
+    border-radius: 0.5em;
+  }
+
+  .form-label {
+    font-size: clamp(1rem, 1.4vw, 3rem);
+  }
+
+  .form-submit {
+    font-size: clamp(1rem, 1.4vw, 3rem);
+  }
+
+  #email-message {
+    top: 32%;
+  }
+
+  .contact-form {
+    min-width: 80%;
+    max-width: 80%;
+  }
+
+  #email-message {
+    top: 36%;
+  }
+
+  .submit-confirmation {
+    top: -10%;
+    left: 20%;
+  }
+}
