+++ src/app/components/GlowingCursor/GlowingCursor.css
@@ -1,0 +1,9 @@
+.glowing-cursor {
+  position: fixed;
+  width: 1px;
+  height: 1px;
+  box-shadow: 0px 0px 180px 100px rgba(197, 152, 113, 0.085);
+  border-radius: 50%;
+  pointer-events: none;
+  z-index: -1;
+}
