+++ src/app/components/Header/Header.css
@@ -1,0 +1,98 @@
+#loading-screen {
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+#welcome-txt {
+  text-align: center;
+}
+
+#top {
+  align-items: flex-start;
+  display: flex;
+  flex-direction: column;
+  position: relative;
+  transform-style: preserve-3d;
+  margin: 0 10% 0;
+}
+
+.parallax-boops {
+  width: 100%;
+  height: 100%;
+  position: absolute;
+  top: -5%;
+  left: 0;
+  z-index: -1;
+  transition: all 0.9s;
+}
+
+#parallax-top-boops {
+  -ms-transform: translateZ(-10px) scale(2);
+  -webkit-transform: translateZ(-10px) scale(2);
+  -moz-transform: translateZ(-10px) scale(2);
+  -o-transform: translateZ(-10px) scale(2);
+  transform: translateZ(-10px) scale(2);
+}
+
+#parallax-top-tendrils1 {
+  -ms-transform: translateZ(-15px) scale(2.6);
+  -webkit-transform: translateZ(-15px) scale(2.6);
+  -moz-transform: translateZ(-15px) scale(2.6);
+  -o-transform: translateZ(-15px) scale(2.6);
+  transform: translateZ(-15px) scale(2.6);
+}
+
+#parallax-top-tendrils2 {
+  -ms-transform: translateZ(-17px) scale(2.8);
+  -webkit-transform: translateZ(-17px) scale(2.8);
+  -moz-transform: translateZ(-17px) scale(2.8);
+  -o-transform: translateZ(-17px) scale(2.8);
+  transform: translateZ(-17px) scale(2.8);
+}
+
+#name {
+  width: 70%;
+  margin-top: 15%;
+  transform-style: preserve-3d;
+}
+
+h1 {
+  font-family: 'Koolegant', sans-serif;
+  font-size: clamp(3rem, 12vw, 10rem);
+  letter-spacing: 0.2em;
+  margin: 0;
+}
+
+h2 {
+  font-size: clamp(1.5rem, 4vw, 5rem);
+  margin: 0;
+}
+
+#mid-bio {
+  font-size: clamp(0.8rem, 3vw, 4rem);
+  margin: 0;
+}
+
+
+/* ------ Media Queries ------- */
+
+
+@media screen and (max-width: 768px) {
+  #top {
+    margin: 0 5% 0;
+  }
+
+  #name {
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+  }
+
+  #mid-bio {
+    font-size: clamp(1rem, 5vw, 3rem);
+    margin-top: 0;
+  }
+}
