+++ src/app/components/Nav/Nav.css
@@ -1,0 +1,45 @@
+#nav-container {
+  display: flex;
+  flex-direction: row-reverse;
+  justify-content: space-between;
+  position: sticky;
+  margin: 3%;
+  top: 5%;
+  z-index: 2;
+}
+
+nav {
+  display: flex;
+  flex-direction: column;
+}
+
+.nav-link {
+  margin-bottom: 3%;
+  text-align: right;
+}
+
+a {
+  text-decoration: none;
+  color: var(--beige-1);
+}
+
+a:hover {
+  color: var(--white);
+}
+
+#socials-container {
+  display: flex;
+  flex-direction: column;
+  font-size: clamp(0.8rem, 2.5vw, 3rem);
+  justify-content: flex-end;
+}
+
+/* ------ Media Queries ------- */
+
+
+@media screen and (max-width: 768px) {
+  nav,
+  #socials-container {
+    display: none;
+  }
+}
