+++ src/app/main.css
@@ -1,0 +1,167 @@
+@font-face {
+  font-family: 'Koolegant';
+  src: url('/fonts/Koolegant.ttf') format('truetype');
+  font-weight: normal;
+  font-style: normal;
+}
+
+@font-face {
+  font-family: 'Satoshi';
+  src: url('/fonts/Satoshi.otf') format('opentype');
+  font-weight: normal;
+  font-style: normal;
+}
+
+*,
+html {
+  scroll-behavior: smooth !important;
+}
+
+:root {
+  --white: #F6F2F0;
+  --beige-1: #a89c90;
+  --beige-2: #988570;
+  --beige-3: #625446;
+  --brown: #4a3f33;
+  --brown-2: #2e2c29;
+  --brown-3: #3f3c39;
+  --dark-brown-1: #1d1c1b;
+  --dark-brown-2: #292520;
+  --black-1: #0e0e0e;
+  --black-2: #131212;
+  --black-3: #0b0b0a;
+  --black-4: #111111;
+}
+
+main {
+  position: relative;
+}
+
+body {
+  background: linear-gradient(75deg, var(--black-1), var(--black-2),
+      var(--black-3), var(--dark-brown-1), var(--black-4));
+  background-size: 300% 300%;
+  animation: gradient 30s ease infinite;
+  margin: 0;
+  overscroll-behavior: none;
+  color: var(--white);
+  font-family: 'Satoshi', sans-serif;
+  font-size: clamp(1rem, 1.8vw, 1.4rem);
+  letter-spacing: 0.15em;
+  height: 100vh;
+  overflow-x: hidden;
+  transition: all 0.9s;
+}
+
+#parallax-wrapper {
+  overflow-y: visible;
+  overflow-x: hidden;
+  perspective: 15px;
+  height: 100vh;
+  position: relative;
+}
+
+
+/* ------ Content Areas ------- */
+
+
+.content-area {
+  margin: 5% 20% 10%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  line-height: 1.5em;
+}
+
+h3 {
+  font-family: 'Koolegant', sans-serif;
+  font-size: clamp(1.4rem, 4vw, 3rem);
+  letter-spacing: 0.2em;
+}
+
+.about-me-p {
+  margin-bottom: 3%;
+}
+
+
+/* ------ Media Queries ------- */
+
+
+@media screen and (max-width: 1205px) {
+  .parallax-boops {
+    top: -15%;
+  }
+}
+
+@media screen and (max-width: 1105px) {
+  h4 {
+    margin: 3% 0 3%;
+  }
+
+  .parallax-boops {
+    top: -35%;
+  }
+}
+
+@media screen and (max-width: 882px) {
+  .parallax-boops {
+    top: -60%;
+  }
+}
+
+@media screen and (max-width: 768px) {
+  body {
+    font-size: clamp(1rem, 3vw, 3rem);
+    letter-spacing: 0.12em;
+  }
+
+  h1 {
+    font-size: clamp(5rem, 19vw, 10rem);
+  }
+
+  h2 {
+    font-size: clamp(2rem, 8vw, 4rem)
+  }
+
+  h3 {
+    font-size: clamp(2rem, 5vw, 3rem);
+    margin: 10% 0 5%;
+  }
+
+  h4 {
+    margin: 3% 0 3%;
+  }
+
+  .mobile-hide {
+    display: none;
+  }
+
+  .parallax-boops {
+    top: -145%;
+  }
+
+  .content-area {
+    margin: 10% 5% 10%;
+    letter-spacing: 0.14em;
+    line-height: 1.5em;
+  }
+
+  .glowing-cursor {
+    visibility: hidden;
+  }
+}
+
+
+@keyframes gradient {
+  0% {
+    background-position: 0% 50%;
+  }
+
+  50% {
+    background-position: 100% 50%;
+  }
+
+  100% {
+    background-position: 0% 50%;
+  }
+}
